refactor: Order -> Orders 클래스명 변경 및 엔티티 설정, setter 삭제
