+++ src/main/java/kitchenpos/domain/Orders.java
@@ -1,0 +1,46 @@
+package kitchenpos.domain;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Entity
+public class Orders {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @ManyToOne(fetch = FetchType.LAZY)
+    private OrderTable orderTable;
+    @Enumerated(value = EnumType.STRING)
+    private OrderStatus orderStatus;
+    private LocalDateTime orderedTime;
+
+    public Orders() {
+    }
+
+    public Orders(Long id, OrderTable orderTable, OrderStatus orderStatus, LocalDateTime orderedTime) {
+        this.id = id;
+        this.orderTable = orderTable;
+        this.orderStatus = orderStatus;
+        this.orderedTime = orderedTime;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public OrderTable getOrderTable() {
+        return orderTable;
+    }
+
+    public OrderStatus getOrderStatus() {
+        return orderStatus;
+    }
+
+    public LocalDateTime getOrderedTime() {
+        return orderedTime;
+    }
+
+    public boolean isCompleted() {
+        return orderStatus.equals(OrderStatus.COMPLETION);
+    }
+}
